backend/requirements.txt
# Core framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Pydantic v2 + settings
pydantic==2.9.2
pydantic-settings==2.5.2

# Utilities
httpx==0.27.2
loguru==0.7.2
python-dotenv==1.0.1
aiofiles==24.1.0

# Database
pymongo[srv]==4.9.2
motor==3.6.0

# Embeddings + ML
sentence-transformers==3.1.1
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3

# LangChain ecosystem
langchain==0.3.3
langchain-openai==0.2.2
langgraph==0.2.26

# Transcript loaders
youtube-transcript-api==0.6.2
webvtt-py==0.5.1
srt==3.5.3

# LLM clients
openai==1.51.2
anthropic==0.37.1

# Audio transcription (Whisper + deps)
openai-whisper==20231117
ffmpeg-python==0.2.0
yt-dlp==2025.1.15   # actively maintained YouTube downloader

# Testing
pytest==8.3.3
